Source code (runnable py)
As noted in the report, I've worked using a Collaboratoy .ipynb notebook and have been reading and writing the files to my Drive folder.
Still, I think this .py should work to read "test_i.txt" from the same directory as the .py is and to create "test_o.txt" there as well. 
No arguments are needed as the names of the files are hardcoded in the program (but can be easily found and changed if the paths weren't the appropriate ones). 
If this doesn't work to rerun the code and replicate the results on your machine (because of some unexpected exception or issues with some of the libraries specified in the report) feel free to reach out to me about it or try to run online the executable sections of the .ipynb that I'll also have uploaded to this repository.